AI_Backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
opencv-python==4.8.1.78
numpy==1.26.4
Pillow==9.5.0
python-multipart==0.0.6
